+++ example/index.js
@@ -4,8 +4,21 @@
 const Pebble = require('..')
 
 async function main() {
-  const pebbleProcess = await Pebble.spawn()
-  pebbleProcess.kill()
+  console.log('\n⏳ Launching Pebble server…\n')
+
+  await Pebble.ready()
+
+  console.log('✔ Pebble server launched and ready.')
+  console.log('✔ Node.js’s TLS module patched to accept Pebble’s CA certificates.')
+
+  // Do stuff that requires Pebble here.
+  // …
+
+  console.log('\n⏳ Shutting down Pebble server…\n')
+
+  await Pebble.shutdown()
+
+  console.log('✔ Pebble server shut down.\n')
 }
 
 main()