+++ index.js
@@ -8,9 +8,11 @@
  * @copyright © 2020 Aral Balkan, Small Technology Foundation
  * @license AGPL version 3.0 or later
  */
-const os           = require('os')
-const path         = require('path')
-const childProcess = require('child_process')
+const os             = require('os')
+const path           = require('path')
+const childProcess   = require('child_process')
+const log            = require('./lib/util/log')
+const MonkeyPatchTls = require('./lib/MonkeyPatchTls')
 
 const spawn = childProcess.spawn
 
@@ -24,9 +26,15 @@ class Pebble {
    * @static
    * @args {[String[]]|String} Optional space-delimited list or array of arguments to pass to Pebble process.
    * @env {Object={ PEBBLE_VA_NOSLEEP: 1, PEBBLE_WFE_NONCEREJECT: 0 }} Optional  environment variables to set for Pebble process.
-   * @returns {Promise<ChildProcess>} Promise to return spawned child process. You’re responsible for managing its life-cycle.
+   * @returns {Promise<ChildProcess>} Promise to return spawned child process.
    */
-  static spawn (args = [], env = { PEBBLE_VA_NOSLEEP: 1, PEBBLE_WFE_NONCEREJECT: 0 }) {
+  static async ready (args = [], env = { PEBBLE_VA_NOSLEEP: 1, PEBBLE_WFE_NONCEREJECT: 0 }) {
+
+    if (this.#pebbleProcess !== null) {
+      // Existing process exists, return that.
+      return this.#pebbleProcess
+    }
+
     // Spawn expects argument to be an array. Automatically convert a space-delimited arguments string to one.
     if (typeof args === 'string') {
       args = args.split(' ')
@@ -41,24 +49,63 @@ class Pebble {
       cwd: binDirectoryPath
     }
 
-    const pebbleProcess = spawn(pebbleBinaryPath, args, options)
+    this.#pebbleProcess = spawn(pebbleBinaryPath, args, options)
 
     let output = ''
 
-    return new Promise((resolve, reject) => {
+    await new Promise((resolve, reject) => {
       const timeoutInterval = setTimeout(() => {
         reject('Timed out while attempting to spawn Pebble server (waited 5 seconds).')
       }, 5000)
 
-      pebbleProcess.stdout.on('data', data => {
+      this.#pebbleProcess.stdout.on('data', async data => {
         output = `${output}${data}`
         if (output.includes('ACME directory available at: https://0.0.0.0:14000/dir')) {
+          //
+          // Pebble server is running.
+          //
+          log(output)
           clearInterval(timeoutInterval)
-          resolve(pebbleProcess)
+          resolve()
         }
       })
     })
+
+    //
+    // To access Pebble from Node (from unit tests, for example), Node needs to accept
+    // Pebble’s certificates:
+    //
+    //   1. Pebble test certificate           : (static; from disk) to access the Pebble server itself over HTTPS.
+    //   2. Pebble root CA certificate        : (dynamically generated) to access servers certified by Pebble.
+    //   3. Pebble intermediary CA certificate: (dynamically generated) to access servers certified by Pebble.
+    //
+    // We have to patch TLS twice to allow seamless support for Pebble certificates in Node.js.
+    // (The first time to access the Pebble server itself over HTTPS and the second time after we’ve downloaded
+    // the dynamically-generated Pebble certificates.)
+    //
+    await MonkeyPatchTls.toAcceptAllPebbleCertificates()
+
+    return this.#pebbleProcess
   }
+
+  static shutdown() {
+    return new Promise((resolve, reject) => {
+      this.#pebbleProcess.on('close', () => {
+        this.#pebbleProcess = null
+        resolve()
+      })
+      this.#pebbleProcess.kill()
+    })
+  }
+
+  //
+  // Private
+  //
+
+  /**
+   * @type ChildProcess
+  */
+  static #pebbleProcess = null
 }
 
 module.exports = Pebble